+++ vico/core/src/main/java/com/patrykandpatrick/vico/core/cartesian/data/LineCartesianLayerDrawingModel.kt
@@ -27,12 +27,12 @@ import com.patrykandpatrick.vico.core.common.orZero
  * top of the [LineCartesianLayer] as a fraction of the [LineCartesianLayer]’s height.
  */
 public class LineCartesianLayerDrawingModel(
-  private val pointInfo: List<Map<Float, PointInfo>>,
+  private val pointInfo: List<Map<Double, PointInfo>>,
   public val zeroY: Float,
   public val opacity: Float = 1f,
 ) : DrawingModel<LineCartesianLayerDrawingModel.PointInfo>(pointInfo) {
   override fun transform(
-    drawingInfo: List<Map<Float, PointInfo>>,
+    drawingInfo: List<Map<Double, PointInfo>>,
     from: DrawingModel<PointInfo>?,
     fraction: Float,
   ): DrawingModel<PointInfo> {
